fix 2 bugs
Зробив деструктуризацію contactsDataBase  та виніс Schema в окремий файл
