+++ controllers/controllers.js
@@ -1,14 +1,20 @@
-const contactsDataBase = require("../models/contacts");
+const {
+  listContacts,
+  getContactById,
+  createContact,
+  deleteContact,
+  updateContact,
+} = require("../models/contacts");
 const { HttpError } = require("../helpers/index");
 
 async function getContacts(_, res) {
-  const contacts = await contactsDataBase.listContacts();
+  const contacts = await listContacts();
   return res.json(contacts);
 }
 
 async function getContact(req, res, next) {
   const { contactId } = req.params;
-  const contact = await contactsDataBase.getContactById(contactId);
+  const contact = await getContactById(contactId);
   if (!contact) {
     return next(HttpError(404, "Not found"));
   }
@@ -17,17 +23,17 @@ async function getContact(req, res, next) {
 
 async function addContact(req, res) {
   const { name, email, phone } = req.body;
-  const newContact = await contactsDataBase.addContact(name, email, phone);
+  const newContact = await createContact(name, email, phone);
   return res.status(201).json(newContact);
 }
 
 async function removeContact(req, res, next) {
   const { contactId } = req.params;
-  const contact = await contactsDataBase.getContactById(contactId);
+  const contact = await getContactById(contactId);
   if (!contact) {
     return next(HttpError(404, "No contact"));
   }
-  await contactsDataBase.removeContact(contactId);
+  await deleteContact(contactId);
   return res.status(200).json(contact);
 }
 
@@ -37,7 +43,7 @@ async function changeContact(req, res, next) {
   if (!body) {
     return next(HttpError(400, "missing fields"));
   }
-  const contact = await contactsDataBase.updateContact(contactId, req.body);
+  const contact = await updateContact(contactId, req.body);
   if (!contact) {
     return next(HttpError(404, "not found"));
   }
