+++ models/contacts.json
@@ -58,6 +58,11 @@
     "name": "Alec Howard",
     "email": "Donec.elementum@scelerisquescelerisquedui.net",
     "phone": "(748) 206-2688"
+  },
+  {
+    "id": "1fa911ad-42e5-48e6-9d92-e5e01d5e41cc",
+    "name": "Vasyl",
+    "email": "vasyl@mail.com",
+    "phone": "077-77-77"
   }
-
 ]
